@@ -1,13 +1,12 @@
 module github.com/open-policy-agent/opa
 
-go 1.16
+go 1.17
 
 require (
 	github.com/OneOfOne/xxhash v1.2.8
 	github.com/bytecodealliance/wasmtime-go v0.36.0
 	github.com/containerd/containerd v1.6.8
 	github.com/dgraph-io/badger/v3 v3.2103.2
-	github.com/dgryski/go-farm v0.0.0-20200201041132-a6ae2369ad13 // indirect
 	github.com/fortytw2/leaktest v1.3.0
 	github.com/foxcpp/go-mockdns v0.0.0-20210729171921-fb145fc6f897
 	github.com/fsnotify/fsnotify v1.5.4
